Restructure S3 auth and source selection for Imager.
There is now capability for only one S3 region/bucket configuration
in Imager, as requests that need to pull out of alternative sources
are expected to set such information in the 'X-S3-Region' and
'X-S3-Bucket' HTTP headers. Thus, the need for an extra configuration
file has been removed and the variables have been merged in the main
configuration file.

Also clarified the need for access and secret keys, since the host
running Alfred may be authenticated for access via IAM.